instagram_dm_mcp_requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
exceptiongroup==1.3.0
fastmcp==2.8.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
instagrapi==2.1.5
markdown-it-py==3.0.0
mcp==1.9.4
mdurl==0.1.2
openapi-pydantic==0.5.1
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
Pillow>=8.1.1
PySocks==1.7.1
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.4
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.47.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
openai
pillow
textblob
